refactor: relocate walk-actions from app/lib to lib/actions
