+++ web/app/show/[id]/page.tsx
@@ -1,7 +1,7 @@
+import { getItemAction } from '@/lib/actions/walk-actions'
 import ItemFetcher from '@/lib/utils/item-fetcher'
 import { idToShowUrl } from '@/lib/utils/meta-utils'
 import ItemBox from '../../../lib/components/item-box'
-import { getItemAction } from '../../lib/walk-actions'
 
 export default function Page() {
   return (
