+++ web/lib/utils/config.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { createContext, useContext, useEffect, useState } from 'react'
-import { getConfig } from '@/app/lib/walk-actions'
+import { getConfig } from '@/lib/actions/walk-actions'
 import { ConfigT } from '@/types'
 
 const ConfigContext = createContext<ConfigT | null>(null)
