+++ web/lib/utils/searcher.tsx
@@ -1,8 +1,8 @@
 'use client'
 import { useSearchParams } from 'next/navigation'
 import { useActionState, useEffect, useTransition } from 'react'
+import { searchAction } from '@/lib/actions/walk-actions'
 import { DataT } from '@/types'
-import { searchAction } from '../../app/lib/walk-actions'
 import { useConfig } from './config'
 import { useData } from './data-context'
 import { useUserContext } from './user-context'
